Optimize literal, doc comment lint as well, extract function.
